Move Types into index.ts
